+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD
 
+* Remove `word_wrap` dependency and method.
+
 * Add `TWEnv::Command::TwitterQuery`.
 
 * Add `TWEnv::Command::PerformActionOnTweets`.